Add files via upload
This project objective was to predict real estate prices in California, there is pre-processing, visualization, fine-tuning and evaluation methods I used in this project. This was a beginner / intermediate project that was one of my first projects that familiarized me with machine learning. I was gonna clean up the code but I left it to show my progress as a coder and data scientist . I can ensure you the quality of work will improve!